SX: Wait for edited translation to be applied before closing editor
Since, applying the edited translation can take some time, especially
for block templates that need to be re-rendered as HTML, it would make
sense to wait for it to be applied before moving back to "Pick a
sentence" step. In the meanwhile, we should display the waiting screen
(with the robot icon) so that the user doesn't experience weird jumps
inside the "Pick a sentence" step.

At the same time, this patch fixes the initial selection of the
translation segment, when coming back to the "Pick a sentence" step
from the SXEditor, for restored draft translations. Until now, the
first segment after the last translated segment was selected, but this
patch modifies the selection to pick the next segment after the one
that has just been edited, in order to be consistent with the user
experience for non-restored translations.

Bug: T336585
Change-Id: I9bc70888ffd96f7df61bd692d17e3aa52ac6bdc7