CX: Use "isForSaving" mode for proper conversion of sentences
In order to support restoration of CX (desktop) translations to SX
(mobile) translations, we need to store the text in the parallel
corpora, in the form of sentences, that is to keep the "cx-segment"
class that is given to each translation segment by the cxserver.

Nevertheless, this class (as well as the data-segmentid attribute) is
unwanted in other cases, like copying content from the Visual Editor.

For this reason, during the node-to-dom conversion by the VE converter,
we only want to keep this markup when the document is being converted
for saving or for translating.

Until recently, the document conversion for saving was using the
"parser" mode, and thus checking for the parser mode was enough to
successfully identify the "saving" process and convert the sentences
appropriately (preserving the "cx-segment" class). However, a fix for
T278323 changed the mode of the conversion for saving to clipboard,
breaking the conversion of sentences for saving, as the "cx-segment"
class is not anymore preserved along with the rest of the text inside
cx corpora.

In order to fix this regression, this patch adds a new "isForSaving"
property to the VE converter and checks for it (instead of checking
for the parser mode), when converting a sentence element, so that
it properly preserves the "cx-segment" class and the "data-segmentid"
when converting the document for both saving and translating.

Bug: T339897
Change-Id: Ic28ec55ae2514c0abe8c1daaa1b040ee88e14c61