CX: Support restoration of templates translated on mobile
Test scenario for this patch:

1. Open SX. Start a draft lead section translation and translate the
   infobox (or any other template) inside the section (along with
   other text contents). You should find an article (lead section)
   for which the infobox can be translated/adapted to the target
   language.
2. Wait 3 seconds for the draft translation to be saved
3. Open CX and select to restore the previous translation.

Expected result: The translated template should be visible inside the
target page contents.

Disclaimer: Publishing the restored article translation doesn't work
with this patch. This is not because this patch introduces a new bug,
but rather because it doesn't solve the issue about the missing
"data-mw-cx-source" attribute for translation units created in SX.
This bug is fixed in the next patch in this chain (I0e20cddc62f736773f0487740c115b9d448d344e
).

Bug: T332863
Change-Id: I996d957ff4aa612d1a03abcdd46977e3fdb3197a