SX: Add "data-mw-cx-source" attribute to corpora units
CX needs the "mw-cx-source" data attribute to be set on corpora
translation units, in order to restore them properly, otherwise
an "Uncaught Error: Attempting to set unmodified MT without an
MT provider" is thrown.

This patch adds this data attribute to the corpora units before
saving on SX, so that the translations started on mobile can be
restored on desktop without the above error.

Bug: T332863
Change-Id: I0e20cddc62f736773f0487740c115b9d448d344e