SX: Fix handling of "sxsave" response in debounced save translation
Recently we updated the response of the "saveTranslation" method of the
translator api, to return the section translation id on success, but we
missed to update the handling of that response inside the debounced
saved translation method, leading to eternally retried "sxsave"
requests.

This patch fixes this issue by handling the response properly.

Change-Id: Ia337fc1fde8fbf44059243adbd0983a41e471dbb