SX: Fix section title translation for both lead and non-lead sections
The latest fix to support the restoration of the translated section
titles for non-lead sections, broke the restoration of the page title
translations for lead sections. This happens because the target section
title property is also populated for lead sections, with a dummy string
that denotes that this is a lead section translation (__LEAD_SECTION__).

This patch fixes the issue by assigning the proper translation for each
case.

Bug: T336579
Change-Id: Ia143459f448730061b22716a3a17b74423d13f6f