SX: Ask for confirmation when deleting an in-progress translation
This patch adds a confirmation dialog that asks for user confirmation,
when user clicks on the delete (trash) icon for a draft translation.

Additionally, it adds a "minHeight" parameter for MwDialog component,
so that we can control the min height for dialogs like the one added
in this patch, where the dialog body text requires less space than the
minimum and we do not want extra white space at the bottom of the dialog
body.

Bug: T333689
Change-Id: I3e9732a086e4382a6f396a203f38d0548e45fe19