CX: Use sitemapper current wiki language code for target language init
Currently, when target language is not provided as URL parameter and
there is no "cxTargetLanguage" property set in the local storage, CX
uses the "wgContentLanguage" config parameter to initialize the target
language. However, this parameter is known to not be reliable for our
purposes in CX, as it fails for wikis like simple.wikipedia.org, where
the content language is the same as on en.wikipedia.org. That leads to
wrong target language initialization for wikis like the simple wiki.

For this reason, this patch uses the "getCurrentWikiLanguageCode"
method from mw.cx.Sitemapper to initialize the target language with
the proper language code for such wikis.

Bug: T329442
Change-Id: I5f328eada096a34a762bed4aaff6634b22a3bb00