SX: Handle session expiration
This patch adds a login dialog that should be displayed to the user,
when it is detected that their session has expired. The detection is
happening by checking if user is logged in based on the
'visibilitychange' event, and specifically the 'visibilityState' is
'visible'. Basically, this means that we only check if the user is
logged in, when the user returns to the SX tab, after having left it
for a while.

These checks only happen for installations where Section Translation
is NOT allowed for anonymous users.

Bug: T308836
Change-Id: I053dc5f05b796e7b84831c6c0acc4c4d71d7b891