CX: Create SectionTitleFetcher service and use it
This patch creates a new service called "SectionTitleFetcher", registers
it inside "ServiceWiring.php" and uses it to fetch the target section
titles and their corresponding section numbers, that are needed in order
to properly calculate the position of a newly published section, within
the target article.

This service uses the "parse" action of the Action API endpoint, in
order to fetch the required information, instead of the deprecated
mobile content API (/page/mobile-sections/{title}).

Tests are added for the new service too.

Bug: T337719
Change-Id: I3b51b954390cd4df191451754f4315269a31080c