SX: Improve usability on desktop/large windows
When accessing the mobile version of test Wikipedia, one will notice
that platform will still redirect to CX when starting a new translation
while still keeping the '.m.' part of the URL. This part of the URL
should be useful in determining whether to preserve SX on a large
screen.

This patch enables that feature by checking for isMobileDomain() from
the siteMapper in mediawikiHelper.

Bug: T337917
Change-Id: I241bec3a1715a74a08ca12e3bb6ba398aaeac155