Remove VRS related code from CX
Since T320529 has now been resolved and parser cache filled with
parsoid output, we can now retire VRS logic from the extension and
start making use of the direct mode by default.

A config patch has been made to flip the switch for CX and this
patch depends on it.

Bug: T339228
Depends-On: If7b3a177c653723745c25e1c8ee13114639ba1b1
Change-Id: Ice2b29024aa2e29e39f64a84a7f3432d61f3fbb1