Do not add "ext.cx.contributions" when Contribute entrypoint enabled
The Contribute Special page is increasingly being deployed in production
wikis. For the wikis that have this entrypoint already enabled, we would
like to remove the "New contribution" menu that is added by Content
Translation inside the contributions page.

This patch checks if the special contribute page is enabled for this
wiki and skin, and adds the "ext.cx.contributions" module only when
that special page is not enabled.

Bug: T337294
Change-Id: If8a9f6ff3e303c99d3d00d6799eeb19316007036