Don't show invite to translate for new article when previewing contents
When a user open the edit page to create a new article, we display an
invite to translate the article instead of creating it from scratch.
However, we only want to display the invitation when the edit page is
NOT in preview mode.

This patch checks if the edit page is in preview mode, before adding
the "ext.cx.entrypoints.newbytranslation" module, to avoid displaying
the invitation in this context.

Bug: T330613
Change-Id: I202e762c12f3442f0f78c9fe92e66417c4d74f74