SX: Skip quick tutorial when user already has section translations
Until now, the quick tutorial step is shown for every single
translation, which adds some friction as shown in previous research.

This patch adds "userHasSectionTranslations" getter and uses it to
control when the quick tutorial step should be displayed based on the
following simple logic:
The quick tutorial should be displayed only for users that haven't
translated any section before. When the user has already created a
section translation, even draft, the quick tutorial will be skipped
and the user will be navigated directly to the "Pick a sentence step".

Bug: T279077
Change-Id: Ia70a86e725ba9a5bc09a960af38534ce3b748b35