SX: Separate draft and published translation fetching
Until now, there was a single approach for fetching both draft and
published translations for Section Translation. More specifically:
- there was a single method ("findSectionTranslationsByUser") for
  fetching all translations from the database in the backend
- there was a single DTO for encapsulating the payload for a
  translation in the backend
- there was a single vuex action that sent a request to the backend
  and stored the returned translations inside the vuex state

This patch changes the way that translations are fetched, so that draft
and published translation fetching is now separated. That means that
now, two different requests (one for each status) are needed to fetch
the draft and the published translations. Additionally, there are two
different methods that fetch the draft/published translations from the
database and two different DTO classes to encapsulate draft and published
translation payloads.

The reasoning behind this change is that splitting translation fetching
allows us to return an array of objects in the same format from the API
(e.g. DraftTranslationDTO objects) instead of more generic objects, and
thus we can better store and use them in the frontend.

Change-Id: I71ca39d77bdd7ef3b766a0d53946234626542b13