SX: Include missing key in instrumentation request
The key `content_translation_session_position` was missing in
instrumentation requests for SX. This resulted in some missing
instrumentation results as the request to log an event failed.

Since this key is required across all logging events, it has now been
included as a default in the main logEvent method. We keep track of the
key using the localStorage. Each key should be and is unique to a
session. Keys in the localStorage are cleared if the sessionId changes.

Bug: T231316
Change-Id: I51e2f59aa763ef3ba9de92a45b5aee5300297da3