SX: Scroll to selected segment when returning from mobile editor
Currently, when user comes back to "Pick a sentence" step after
editing the proposed translation, the contents are not scrolled
at all, even when the selected translation segment is outside
the viewport. This issue is caused because the watcher that is
called when the selected translation unit is updated, and scrolls
to the newly selected translation unit, is not called when the
"Pick a sentence" step is mounted.

This patch fixes the issue by setting the "immediate" property
for the watcher and adding it inside the "onMounted" hook, so
that the watcher is executed as soon as the step is mounted.

Bug: T341685
Change-Id: I3e1d1570cf58330011d5dc1bd8a58f11fe470771