SX: Fix regression from making quick tutorial step optional
Until recently, the quick tutorial step was always displayed before
moving to "Pick a sentence" step. While the quick tutorial step was
displayed, the application was resolving the references inside page
contents in the background, and there was enough time for this process
to be completed until the user was navigated to the "Pick a sentence"
step.

However, recently the logic changed and the quick tutorial is skipped
for users that already have at least one section translation (draft
or published). That led to less time between confirming a translation
inside "Confirm translation" (for lead section translations) and moving
to "Pick a sentence" step, and now the references are resolved after
the step has been loaded.

Because of the way that Vue reactivity works and because the page
sections with resolved references are set in a way that doesn't trigger
updates in computed variables like the one inside "useSubSectionContent"
composable, the section contents are not anymore reactive leading to
bugs in selecting translation segments and applying the proposed/edited
translation. These bugs are demonstrated in the screencast below:
https://imgur.com/a/IDQYfZK

This patch changes the way that page sections are updated once the
references are resolved, so that the reactivity for the page contents
is preserved and the above bugs are resolved.

Patch where skipping quick tutorial was added:
Ia70a86e725ba9a5bc09a960af38534ce3b748b35

Change-Id: I8651529fe94953038be6d1d076c7ef9cc05fb791