SX: Fix draft restoration after having started another SX
When going back to dashboard from "Pick a sentence" step, the current
section suggestion vuex state variable is not cleared. Until recently,
when we didn't support draft translation restoration, this was not an
issue, because when selecting a new page or section to translate, this
state variable ("currentSectionSuggestion") is set to a new value.
However, restoring a draft translation doesn't set the current section
suggestion, and thus the "currentSectionSuggestion" variable holds the
previously set value (if any) leading to wrong content being displayed
to the user after restoration.

To fix this issue, this patch clears the "currentSectionSuggestion"
state variable when going back to dashboard. Additionally, it also
clears the "currentSourceSection" state variable (even if it doesn't
produce such issues), so that the state doesn't contain any outdated
information after the user has left a translation.

Bug: T341765
Change-Id: Ib1dffe57737eb64e743bd054ebab35c50980b254