CX UI lib: Minor fixes in MWSpinner component styles
This patch removes the margin-right values that are used for the "before"
and "after" pseudo-elements of this component. These values are not
needed and were invalid anyway, since they were set to "1em / 4".

Additionally, this patch sets the width of the spinner to
3 * @width-bounce-dot, since there are exactly 3 dots, included in the
spinner, so its width should match its contents.

Change-Id: Ied031aeedda6fdb5fb0efdcaebf0e0328e956ced