CX: Refactor mw.cx.SiteMapper.js
This patch refactors mw.cx.SiteMapper.js file to use ES6 instead of
ES5. Additionally, it removes the usage of jQuery in this file.

Since this produces some replacements of jQuery promises with regular
ES6 promises, the patch also fixes the usage of the methods that have
their returned value changes. Basically this refers to the
"getLanguagePairs" method which used to return a jQuery promise and
now it returns a regular promise, so we also fix its usage inside
"ext.cx.dashboard.js" and "site.js" files.

Bug: T342280
Change-Id: I2c288c99891717304fc2bfd9e56998aaaf144085