CX: Pre-select target language in Uls Entrypoint
* Hot fix
When selecting the `...` in the ULS entry point under 'missing in other
languages', the target language is not set as we don't know what
language the translator would want to translate to. This prevents the
SelectedSourcePage in Special:ContentTranslation from showing up.

This patch selects the first language from the suggested languages to
translate to and sets it as the target language.

Bug: T311820
Change-Id: I7829fded49a18c2d70a44ff97b2197918e6a39cd