SX: Wrap buttons on SXConfirmBackNavigationDialog
In en, the two buttons; Continue editing and Discard changes, may appear
on the same row due to the fact that the text can fit. In other
languages such as Greek, the translated text may appear way longer and
break the layout.

This patch fixes the issue by adding wrap styling for the dialog
component.

Bug: T341193
Change-Id: Ibe2d52e3f65635e5f12a887ed816a5886dc367ab