SX: Add useDevice composable and use it where needed
Now that we are working towards adopting the SX dashboard as the unified
dashboard, we have to follow different approaches for different devices
on several occasions.

This patch adds the "useDevice" composable which returns computed
properties (only one for now) that help identify the device and act
appropriately.

There are already two usecases where this composable is needed and more
will be added soon.

Change-Id: Ie38d184eb24992bae1f783dd151e000330a19b28