SX: Add useCXRedirect composable and use it where needed
Once the SX dashboard is used as the unified CX/SX dashboard, we'll
need to redirect to CX translation view, when the user starts a new
translation on desktop or resumes a draft translation on desktop.

This patch adds the "useCXRedirect" composable that returns a method
that handles the CX token setting and the redirection to CX.

This patch also uses the above composable where needed. The same
composable will also be used in subsequent patches.

Change-Id: I391f9dca7f55f79359d42e71c62e889d435fdc48