SX: Fix resuming draft translation started with different language pair
When the user accesses the mobile dashboard in "X" language to continue
on mobile a translation for an article from English to "Y" (that was
just started on desktop), the desktop dashboard is loaded on mobile
with the confirmation step, instead of the user being redirected to
the target wiki and the draft translation being loaded there.

This happens because the translation was just created in CX and thus
the CX token for the given source language, target language and source
title is set inside a cookie. However, this cookie is used inside
SpecialContentTranslation.php file to determine if the dashboard or
the CX translation view should be displayed. Finally, this token
leads to unexpected results, when the user tries to restore a draft
translation started with a different target language than the current
one.

This patch solves this issue by unsetting this token for these cases,
when such a draft translation is clicked.

Bug: T343706
Change-Id: I837b6d046e8e57b4d11496cadbc715e005bd22d5