SX: Wrap buttons on SXConfirmTranslationDeletionDialog
In en, the two buttons; Cancel and Delete Translation may appear
on the same row due to the fact that the text can fit. In other
languages such as Greek, the translated text may appear way longer and
break the layout.

This patch fixes the issue by adding wrap styling for the dialog
component.

Change-Id: I2936cbba041ebfe8e0aa432c4a26f82fc8fe38dc