SX: Wait for "Pick a sentence" step to be loaded before scrolling
Currently, when a draft translation is loaded inside the
"Pick a sentence" step, the contents of the page are scrolled to the
currently selected translation segment, once the proper translation
unit is selected. However, this selection can (and it does) happen
before the contents of the section have been rendered, leading to
an error.

This patch adds a waiting of the next tick, before selecting the
initial translation segment. This way, we make sure that the section
contents have been rendered inside the page, before scrolling to the
selected segment.

Change-Id: I98c73ebdbf3ff919afa1df18c750dd121a0256be