SX: Move MT provider initialization functionality to composable
This patch moves the MT provider fetching and initialization
functionality from vuex actions to a composable. The reasoning behind
this move, is that this way we avoid having global vuex actions for a
functionality that is specific to a single SFC, but instead we use a
standalone composable to introduce this functionality where needed in
an isolated way.

The implementation is identical to the previous one with one
important difference: it increases the "translationDataLoadingCounter"
state variable before fetching and initializing the MT providers, so
that we make sure these integral data have been loaded before the
section translation can start inside the "Pick a sentence" step.
The counter is decreased once the initialization has been completed.

This change fixes a regression that has been introduced in I16015126269216b43e3505e2868cf97975982909
and is demonstrated in the screencasts below.

Screencast before increasing the counter during MT provider fetching:
https://phabricator.wikimedia.org/F37477245

Screencast after this patch:
https://phabricator.wikimedia.org/F37477265

Bug: T342530
Change-Id: I37e2fc939e8d91abb4cb37cbce4ac6af0b0d6d70