SX: Check if target is of Element type before accessing classList
Currently, when the user clicks somewhere in the section contents
inside "Pick a sentence" step, we are checking whether any of the
ancestors of the event target has the "cx-segment" class. However,
when this class is missing from all the ancestors (typically happens
for nested inline templates), the checks goes up all the way to the
root HTMLDocument and Window objects, which (as they are not instances
of "Element" class) do not any "classList" property, and trying to
access this property causes an error.

This patch check if the ancestor is of Element type before accessing
the "classList" property to avoid such errors.

Bug: T323233
Change-Id: I04103afc54b67ffa8c0cc87e6ed93c0c0c551db6