SX: Move resolution of content references from vuex action to composable
This patch moves the resolution of content references from vuex actions
to a composable. The reasoning behind this move, is that this way we
avoid having global vuex actions for a functionality that is not limited
to vuex mutations, but instead we use a standalone composable to
introduce this functionality where needed in an isolated way.

The implementation is identical to the previous one with one
important difference: it includes the loading of the Visual Editor
modules, which are prerequisites for resolving the content references.

This way we also fix an issue that was introduced by the new section
translation suggestions for published translations (inside the dashboard
lists).

Screencast of the issue that is being fixed:
https://imgur.com/a/18Migco

Change-Id: Ie919bc9d9578ae6a438395b260369c5e9ebe9386