SX: Also check status when checking if translation already exists
This patch modifies the "translator/translationExists" vuex getter to
also check the existence of a translation based on its status (draft
or published).

Until now, this getter only checked for the translation id and the
section translation id of the given translation. This led to identifying
as existing, missing draft translations because a published translation
with the same translation/section translation ids already existed in the
store.

In order to properly detect all existing translations and avoid skipping
of storing unsaved translations, also check for the translation status.

Change-Id: I6f882420c36d9921a96eec45b4f84aad54f639ce