SX New Section Placeholder: Receive translation messages as arguments
Currently, inside "Compare the contents" step, when the selected
section to translate is a missing section, and the target article tab
we need to position the new section placeholder, inside the contents
of the target page, just before the first appendix section.

However, in order to render a Vue component programmatically, we need
to create a new Vue app, which misses the "v-i18n" directive. For this
reason, a Vue warning is thrown inside the console:
```
Failed to resolve directive: i18n
  at <SxContentComparatorNewSectionPlaceholder isMappedSection=false i18n=fn<bound i18n> >
```
Screenshot of the warning:
https://phabricator.wikimedia.org/F41596618

In order to avoid this warning, this patch refactors the
NewSectionPlaceholder SFC, to accept the translation messages as props,
instead of computing them inside setup function, based on other props.

Change-Id: I0bff320418532847500aa1af0f799547f21724f7