SX: Assert user before each router navigation
When a logged-in user visits SX application, we need to make sure that
their session is not expired, at any step of the publishing workflow,
in order to avoid the published contents to get associated to the user
IP instead of the user profile without any notice.

Additionally, in order to avoid losing any kind of work by the user
(section selection, contents comparison, or even worse translation),
this patch adds user assertion checks before each router navigation,
so that we catch expired sessions as soon as possible, without having
to send requests every x seconds.

Once the user assertion is also added for "sxsave" (autosave) and
"cxpublishsection" (publishing) requests - these checks will be added
in a following patch -, the handling of expired sessions will be
completed.

Bug: T308836
Change-Id: Ic2d0f5f201769228fce8ba702cb0d3e00101600a