CX: Refactor TranslationController/processSaveQueue method
Until now, the "processSaveQueue" method of the TranslationController
integrates a lot of functionalities related to the auto-saving process.
More specifically, among other things it handles the request preparation
and the success and failure results. However, this increases its size to
more than 140 lines and makes it difficult to read and review.

This patch refactors the method by extracting the save request creation,
and the success and failure handlers to separate methods, for better
separation of concerns and increased readability and maintainability.

Change-Id: I8207bb94b70da671bfda2d90e3f8e6c9c2021ec0