CX: Use "sxsave" action when translating a section on desktop
Until now when user translates a section on desktop, the auto-save
functionality uses the "cxsave" API endpoint. However, this endpoint
is only suited for article translations, while auto-saving for section
translations should use the "sxsave" endpoint.

This patch checks if current translation is a section or an article
translation and sends a request to the appropriate endpoint with the
appropriate payload. It also adds a new "getSaveQueue" method that
returns the proper subsections/paragraphs to be considered for auto
saving, depending on whether it's a section or an article translation.

Bug: T345329
Change-Id: I17aa959ad12b79b689169c1fb2d0b99c609bf258