CX: Add "sourcesection" URL param if needed when continuing a draft
Resuming a draft translation through the unified dashboard in desktop
mode should redirect to the desktop editor. Desktop editor also supports
section translations, but for that to happen, a URL parameter named
"sourcesection" filled with the source section title, should be present.
That's something that is not happening with the current implementation.

This patch properly adds the "sourcesection" URL parameter before
redirecting to the desktop editor to properly support continuing a draft
section translation on desktop.

Bug: T343699
Change-Id: Ieafa465f03550d6076a6401241c72c8ddf277ab8