CX: Add "target_url" to published translation DTO
Published translations always have the target URL set inside the
"cx_translations" table. This URL is needed inside the dashboard,
so that we can properly redirect the user to the published article,
when clicking on a translation inside the published translation list.

This patch adds this "targetUrl" variable to the published translation
DTO classes (both backend and frontend) and makes sure its populated
when these DTOs are created.

Change-Id: Ie6650f23f98086655616a8219b4ba185be05141f