SX unified dashboard: Open target page on published translation click
Currently, when clicking on a published translation inside the CX
dashboard, the user is redirected to the published article. This
behaviour is also expected from the unified dashboard.

This patch adds a click handler for each item inside the published
translation list that redirects to user to the published article,
and changes the translation title element to be an anchor (<a>) for
semantic and accessibility purposes.

Change-Id: Ic1063a8d218ee1fa5524b5bf7387ea8f88bb598a