SX unified dashboard: Enable published translation editing on desktop
Editing published translation is supported in CX dashboard, but there
is not yet a concrete plan about the expected editing behaviour in
mobile mode.

This patch enables editing published translation in unified dashboard,
only for desktop mode, so that we continue to support this behaviour,
even after the dashboard migration. For mobile devices, we'll enable
the editing functionality, once it has been designed.

Change-Id: I06ff6e547d7fee578e752ec9698417e8a5049e4d