SX dashboard: Do not store suggestions for published translations
Currently, the suggestions for page and section translations inside the
SX dashboard, are populated from the user published translations, if
such translations exist.

This is a side-effect of the implementation for the "section expansion
invite" buttons that are included inside the published translation
items. In order to check for suggested sections to expand for each
published translation, we use the "loadSectionSuggestion" vuex action,
which apart from fetching section suggestions for an article (which is
desired), it also stores new suggestions inside the vuex store for
these published articles (which is not desired).

This patch fixes this issue by avoiding using any vuex actions with
complex behaviours, and it just fetches the section suggestions for
each published translation, without storing any suggestion inside the
store.

Bug: T347204
Change-Id: Iddf1353875fa203b127f39a99c6be41271661a94