SX: Scroll to selected segment on selected translation unit update
Recently, a patch that fixed scrolling to the selected translation
unit after returning to "Pick a sentence" step from the mobile VE,
moved the "segment scrolling" functionality from a watcher of the
"selectedContentTranslationUnit" computed variable, to a watcher of
the "translationDataReady" computed variable. However, this move
caused a regression, as section contents are not scrolled anymore
to the selected segment when the newly selected segment is outside
the viewport.

This "segment scrolling" code is necessary to be run on both occasions:
a. upon return to "Pick a sentence" step from the visual editor
b. upon update of the selected translation unit inside "Pick a sentence"

This patch moves all the shared "scroll to the selected content
translation unit" to the related "useSegmentScroll" composable and
renames it to "useSelectedContentTranslationUnitScroll", and it
restores the watcher on the "selectedContentTranslationUnit" computed
variable, so that contents are properly scrolled upon selected
translation unit change.

Notes for testing this patch:
1. When applying or skipping the proposed translation for a sentence
at the edge of the viewport, the contents should be scrolled, so that
the newly selected sentence (next sentence) is visible. (T346027)
2. When the user edits the translation for a segment that is outside
the initial viewport, and then returns to the "Pick a sentence" step,
the contents should be scrolled, so that the selected translation unit
is visible. (T341685)

Bug: T346027
Change-Id: I989633dfdcfe9df73a4611e0c4e0379366e6bf53