SX: Check if user is anon, instead of the related config parameter
Currently, in order to enable the checks for the user session, we check
the "ContentTranslationEnableAnonSectionTranslation" configuration
parameter and if it's true, we enable the user session expiration check.

However, this check is not enough, as even if the SX application is
enabled for anon users, we still need to check the session status for
users that visited the application as logged-in users, in order to avoid
unexpectedly assigning work to the user's IP instead of its account.

For this reason, this patch replaces the check for the config parameter,
with checking if the user visits SX as a logged-in user. This way, we
always check the status of the session for logged-in users, regardless
if anon users are enabled for SX. If they are not, they won't be able
to visit SX, anyway.

Change-Id: I7616d48405aa32e66c67f724807bc7f5a152eb05