SX: Move isLoginDialogOn variable to application state
Until now, the visibility state of the SX login dialog is controlled
by a ref variable set inside the root App.vue file. However, this
implementation prevents the dialog to be opened/closed from other
places.

This patch moves the "isLoginDialogOn" variable from the App.vue
file to the vuex store, so that it is globally available for both
read and mutation. This will enable us to implement the session
expiration check, without depending on the context of the App.vue
SFC.

Change-Id: Ifee8d0f460700ceaa4012299d819557da8c5685a