CX UI lib: Add zero col-offset for each breakpoint to grid classes
This patch adds classes that reset the offset to 0, one for each
breakpoint ("mobile", "tablet", "desktop" and "desktop-wide"). These
classes basically set the "margin-inline-start: 0;" when the
corresponding media query is matched. The 4 classes that are actually
added, can be directly used inside SX application.

Their purpose is to enable the developer to reset an offset set for a
"smaller" breakpoint. For example, if "col-offset-tablet-6" is present
and we want to remove this offset for "desktop" and larger screens, we
can add the "col-offset-desktop-0" class.

These 4 classes are:
"col-offset-mobile-0", "col-offset-tablet-0", "col-offset-desktop-0",
"col-offset-desktop-wide-0".

Change-Id: Iff7c2b7b15e8720c829829d1f7948d5716135fcb