CX publish section: Add section translation id to payload
Currently, publishing a section translation on desktop is blocked, as
the "cxpublishsection" requests are failing. This happens, because
a mandatory parameter is missing from the payload.

This patch stores the id of the current section translation, as
returned by the "sxsave" response, and adds it as the missing
"sectiontranslationid" parameter to the payload of the
"cxpublishsection" requests, to fix this issue.

Bug: T347720
Change-Id: I3a0bd7bc52b5d1f6febb30ed288c4fbdfab514d2