CX section publishing: Auto-save translation before publishing
Currently, the section translation id for a section translation on
desktop, is only set after a successful auto-save request (the section
translation id is returned by the autosave API).

That means, if a user continues a draft translation on desktop, and
publishes it right away, without making any modifications (and thus no
auto-save requests will be sent), the section translation id is unset.
This leads to a publishing error, since "sectiontranslationid" is a
required parameter for the "cxpublishsection" API.

To avoid this blocking issue, this patch always sends an "auto-save"
request, before section publishing on desktop, to retrieve and store
the section translation id, before trying to publish.

This is a quick, temporary to unblock section publishing on desktop for
these cases. Ultimately, we would like to fetch the section translation
id on translation initialization (as part of the "fetchDraftInformation"
response), to avoid these unnecessary auto-save requests before
publishing.

Bug: T347720
Change-Id: Ided548ff8b06155e9c9cab0da5cd26b3fa4066ea