Move static method "findForTranslator" to TranslationStore
Until now, inside Translation class lives a static method named
"findForTranslator". This method expects a "TranslationWork" object
and a Translator instance as parameters, and return an instance of
Translation class, if found.

This patch moves this method to the TranslationStore, and renames
it to "findTranslationByUser". This way we can use proper dependency
injection, instead of using static methods and fetching services using
Mediawiki:getInstance(). Additionally, here we remove the usage of
"TranslationWork" value object, which basically just hides the simple
dependencies on title, source and target language of a translation.

Change-Id: If308a0ed06063e16955997b2f47891200aa2b343