CX: Move find method from Translation class to the TranslationStore
This patch moves the static "find" method from the Translation class,
(which is supposed to be a model), to the TranslationStore service,
and replaces all its usages. This will enable us to use proper
dependency injection, instead of static methods with hidden
dependencies.

Ultimately, we would like to move all the static methods that interact
with the database, from Translation class to the TranslationStore.

The "find" method has not yet been actually removed, since it's still
needed inside "getConflictingTranslations" static method of Translation
class. The "getConflictingTranslations" method will also be moved to
the TranslationStore in a following patch, and this is where the "find"
method will be removed for good.

Change-Id: I0338fa3ed2ed1605c075248e88f58e49438a9627