CX translation init: Transform jquery promises to ES promises
This is part of the work for the modernization of the CX codebase,
(that is contents of the /modules directory). This patch refactors
the methods related to translation data fetching, to return ES
promises instead of jquery promises.

Since we will work on these methods in the following patches, we
will also benefit by this refactoring, given that we won't have
to deal with the differences that the jQuery promise API has.

Change-Id: I2303c85f8465ead71edd2b3296e2f8bf3b655bf6