CX contenttranslation list: Add "usecase" parameter
As described more thoroughly in the linked ticket, currently the
"contenttranslation" list of the query action, serves several
scenarios within the CX/SX applications. The API returns different
responses, based on different parameter combinations, without
communicating it clearly to the (potential) API users.

To provide a clearer API, this patch adds the "usecase" parameter,
that will be used to serve the 5 different scenarios, in which
this list is used within CX/SX applications.

The patch also adds the implementation for the "unified-dashboard"
usecase, so that the parameter is more clearly demonstrated. The
"serveUnifiedDashboardTranslations" method is basically a copy of
the current implementation of the "sectiontranslationsonly" case.
The API request for the "sectiontranslationsonly" case is also
modified to use the "usecase" parameter instead.

Bug: T348563
Change-Id: I9c5cee67a3f0df5f144df1679933b89eb968d473