Add "sectionTranslationId" to desktop draft translation DTOs
This patch adds the "sectionTranslationId" property to the
CXDraftTranslationDTO (PHP) and mw.cx.dm.DraftTranslationDTO (JS)
classes. This field will be populated with the id of the draft
section translation, when the user translates a section translation
on desktop. This property will be null for new section translations,
but must be an integer when a draft translation is resumed.

Given that this field will now transfer the section translation id
to the frontend, we can now remove the "autosave" request, that is
always sent before section publishing on desktop - the only purpose
of this request is to fetch the section translation id of the draft
translation, and is a quick hacky fix that should be removed.

Bug: T347720
Change-Id: I6a6a80677867967118e816118bd5f4a3719cbac5